[fix] match_words value is modified after being cached for next iteration
The script saves the value of match_words as s:match_words, caching it
to avoid extra work the next time the script is called. However, match
words was being modified (by appending the escaped value of the
'matchpairs' option) after the caching occurred. Thus, the cache was
dirty but not recognized as such, and the wrong match_words value was
then used on subsequent invocations.

Bug fixed by moving all match_words modifications before cache is
created or checked.